Destructure authorization  actions from index
