+++ src/components/authorization/LoginFormContainer.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import LoginForm from "./AuthForm";
 import { connect } from "react-redux";
-import { login } from "../../store/actions/";
+import { login } from "../../store/actions/authorization";
 
 class LoginFormContainer extends Component {
   state = { login: "", password: "" };
