+++ src/components/authorization/SignupFormContainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { signup } from "../../store/actions";
+import { signup } from "../../store/actions/authorization";
 import SignupForm from "./AuthForm";
 
 class SignupFormContainer extends Component {
